fix(textlang): only written words may authorize a cut
Three findings from review, all cases where the new signature cut fired
on something that is not a reply.

saysSomething scored the RAW text while WordsWritten counted only what a
person typed. A bare 'Subject: Update on the plan and the budget' then
answered zero authored words and still cleared the stopword bar on its
own 'the's — a metadata line authorizing the removal of the whole quoted
thread below it, which is the exact defect the word floor exists to
prevent. Both questions now read the same text, through a new
WrittenLines.

It also had no floor of its own, and the Vietnamese path is decided by
the SHARE of accented runes, so a single word like 'Résumé' read as
Vietnamese on two of its six letters and waved a cut through. A lone
word is not a reply in any language, whatever it is written in.

'please consider the environment' is gone from the footer leads: it
opens 'the environment variables already configured in production' just
as happily, and cutting there discards a real request.

textlang.go crossed the 500-line cap with these additions, so the
reply-boundary half moves to replytext.go: the quote markers, the
attribution lines, and the floors that stop a cut from leaving nothing
to read. textlang.go answers which language this is; replytext.go
answers which part of it is the reply, and the second is the question
that keeps being got wrong.
